Enum serde_pickle::value::HashableValue

pub enum HashableValue {
    None,
    Bool(bool),
    I64(i64),
    Int(BigInt),
    F64(f64),
    Bytes(Vec<u8>),
    String(String),
    Tuple(Vec<HashableValue>),
    FrozenSet(BTreeSet<HashableValue>),
}

Represents all primitive builtin Python values that can be contained in a "hashable" context (i.e., as dictionary keys and set elements).

In Rust, the type is not hashable, since we use B-tree maps and sets instead of the hash variants. To be able to put all Value instances into these B-trees, we implement a consistent ordering between all the possible types (see below).

Variants

None

None

Bool

Boolean

I64

Short integer

Int

Long integer

F64

Float

Bytes

Bytestring

String

Unicode string

Tuple

Tuple

FrozenSet

Frozen (immutable) set

Methods

impl HashableValue

fn to_value(self) -> Value

Convert the value into its non-hashable version. This always works.